Retain values in loop properly
When a value in a loop is requested to be retained for backprop
later, the value will be pushed to a permanently allocated
stack. The computation graph for backprop pops the stack to
grab retained values in a reverse order.

A couple of op were introduced. Unfortunately, they need a very
special attention to schedule properly. For details, see
ScheduleStackPushPop.